#6 Rename header and navbar components to reduce naming confusion
