+++ src/App.js
@@ -5,14 +5,14 @@ import { withRouter } from 'react-router-dom';
 import AppContainer from './containers/AppContainer';
 
 import Routes from './components/Routes';
-import Navbar from './components/Navbar';
+import Header from './components/Header';
 
 const App = appProps => (
   <AppContainer
     {...appProps}
     render={props => (
       <div>
-        <Navbar
+        <Header
           isAuthenticated={props.isAuthenticated}
           currentUsername={props.currentUser}
           submitLogout={props.submitLogout}
