+++ src/components/Navbar.js
@@ -1,21 +1,59 @@
 import React from 'react';
-import { Nav, NavItem } from 'react-bootstrap';
-import { Link } from 'react-router-dom';
-import { LinkContainer } from 'react-router-bootstrap';
-import { GroveNavbar } from './GroveNavbar';
+import PropTypes from 'prop-types';
+import { Navbar as BSNavbar } from 'react-bootstrap';
+import { UserInfo } from './UserInfo';
 
-const brandLink = props => <Link to="/" {...props} />;
-const Navbar = props => (
-  <GroveNavbar title="MarkLogic Grove" {...props} brandLink={brandLink}>
-    <Nav>
-      <LinkContainer exact to="/">
-        <NavItem>Search</NavItem>
-      </LinkContainer>
-      <LinkContainer exact to="/create">
-        <NavItem>Create</NavItem>
-      </LinkContainer>
-    </Nav>
-  </GroveNavbar>
+export const Navbar = ({
+  logo,
+  logoStyle,
+  title,
+  children,
+  withoutUser,
+  ...props
+}) => (
+  <BSNavbar fluid={true}>
+    <BSNavbar.Header>
+      {logo && (
+        <BSNavbar.Brand>
+          <div className="navbar-left">
+            <props.brandLink>
+              <img
+                src={logo}
+                style={logoStyle || { maxWidth: '100px', maxHeight: '45px' }}
+              />
+            </props.brandLink>
+          </div>
+        </BSNavbar.Brand>
+      )}
+      <BSNavbar.Brand>
+        <props.brandLink>
+          <span>{title}</span>
+        </props.brandLink>
+      </BSNavbar.Brand>
+      <BSNavbar.Toggle />
+    </BSNavbar.Header>
+    <BSNavbar.Collapse>
+      {children}
+      {!withoutUser && (
+        <UserInfo
+          isAuthenticated={props.isAuthenticated}
+          currentUsername={props.currentUsername}
+          submitLogout={props.submitLogout}
+          loginPath={props.loginPath}
+        />
+      )}
+    </BSNavbar.Collapse>
+  </BSNavbar>
 );
 
-export default Navbar;
+const defaultBrandLink = props => <a href="/" {...props} />;
+
+Navbar.defaultProps = {
+  brandLink: defaultBrandLink
+};
+
+Navbar.propTypes = {
+  title: PropTypes.string,
+  brandLink: PropTypes.func,
+  withoutUser: PropTypes.bool
+};
