+++ src/components/Navbar.test.js
@@ -1,18 +1,16 @@
 import React from 'react';
 import { shallow, render } from 'enzyme';
 import expect from 'expect';
-import { GroveNavbar } from './GroveNavbar';
+import { Navbar } from './Navbar';
 
-describe('<GroveNavbar />', () => {
+describe('<Navbar />', () => {
   it('renders without crashing', () => {
-    expect(shallow(<GroveNavbar />).length).toEqual(1);
+    expect(shallow(<Navbar />).length).toEqual(1);
   });
 
   it('includes the title', () => {
     // Have to render to to see into a child. Is there a better way?
-    const wrapper = render(
-      <GroveNavbar title="MyNavbarTitle" withoutUser={true} />
-    );
+    const wrapper = render(<Navbar title="MyNavbarTitle" withoutUser={true} />);
     expect(wrapper.text()).toContain('MyNavbarTitle');
   });
